pyproject.toml
[project]
name = "sovereign-ai-lab"
version = "0.1.0"
description = "European Business Data Analyst Agent - IMT BS demo"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langgraph>=0.2.50",
    "langchain-groq>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-experimental>=0.3.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "fpdf2>=2.8.0",
    "python-dotenv>=1.0.0",
    "duckduckgo-search>=6.3.0",
    "streamlit>=1.40.0",
    "requests>=2.32.0",
    "eurostat>=1.1.0",
    "idun-agent-engine>=0.5.3",
]
